ModalPage
Зробив універсальну ModalPage яка приймає title та children. Приклад застосування дивись компонент NeedHelp
