+++ src/components/SideBar/NeedHelp/NeedHelpModal/need-help-modal.module.css
@@ -1,4 +1,4 @@
-.background {
+/* .background {
   position: fixed;
   top: 0;
   left: 0;
@@ -9,9 +9,9 @@
   display: flex;
   justify-content: center;
   align-items: center;
-}
+} */
 
-.modalBox {
+/* .modalBox {
   display: flex;
   flex-direction: column;
   background-color: var(--primary-color-black);
@@ -23,25 +23,24 @@
   max-width: 400px;
   border: 1px solid rgba(190, 219, 176, 0.5);
   box-shadow: 0 4px 16px 0 rgba(22, 22, 22, 0.05);
-}
+} */
 
-.close {
+/* .close {
   position: absolute;
   top: 10px;
   right: 10px;
   cursor: pointer;
   font-size: 20px;
   color: #fff;
-}
+} */
 
-h2 {
-  /* margin-top: 0; */
+/* h2 {
   font-family: var(--font-family);
   font-weight: 500;
   font-size: 18px;
   letter-spacing: -0.02em;
   color: #fff;
-}
+} */
 
 form {
   margin-top: 24px;
