+++ src/pages/ModalPage/modal-page.module.css
@@ -1,0 +1,53 @@
+.background {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.5);
+  z-index: 1000;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.modalPageBox {
+  display: flex;
+  flex-direction: column;
+  background-color: var(--primary-color-black);
+  color: #fff;
+  padding: 24px;
+
+  position: relative;
+  border-radius: 8px;
+  max-width: 400px;
+  border: 1px solid rgba(190, 219, 176, 0.5);
+  box-shadow: 0 4px 16px 0 rgba(22, 22, 22, 0.05);
+}
+
+.close {
+  position: absolute;
+  top: 10px;
+  right: 10px;
+  cursor: pointer;
+  font-size: 20px;
+  color: #fff;
+}
+
+h2 {
+  font-family: var(--font-family);
+  font-weight: 500;
+  font-size: 18px;
+  letter-spacing: -0.02em;
+  color: #fff;
+}
+
+@media only screen and (min-width: 768px) {
+  .modalPageBox {
+    padding: 40px;
+  }
+}
+
+.errorText {
+  color: red;
+}
